feat: async listAssetPropertiesDescription
