+++ packages/dashboard/src/hooks/useAssetDescriptionQueries.ts
@@ -4,8 +4,6 @@ import {
   type AssetProperty,
   type DescribeAssetResponse,
   type IoTSiteWiseClient,
-  AssetPropertySummary,
-  AssetModelPropertySummary,
 } from '@aws-sdk/client-iotsitewise';
 import type {
   SiteWiseAssetQuery,
@@ -13,7 +11,11 @@ import type {
 } from '@iot-app-kit/source-iotsitewise';
 import { useQuery } from '@tanstack/react-query';
 import { useClients } from '~/components/dashboard/clientContext';
-import { listAssetPropertiesWithComposite } from './listAssetPropertiesWithAssetDescription';
+import {
+  createFetchSiteWiseAssetQueryDescription,
+  createListAssetPropertiesMapCacheKey,
+} from '~/data/listAssetPropertiesMap/query';
+import { selectListAssetPropertiesMap } from '~/data/listAssetPropertiesMap/selectData';
 
 export type PropertySummary = {
   propertyId: AssetProperty['id'];
@@ -125,98 +127,21 @@ export const useAssetDescriptionMapQuery = (
   });
 };
 
-// new call beyond this point
-
-const listAssetProperties = (client: IoTSiteWiseClient, assetId: string) => {
-  return new listAssetPropertiesWithComposite({ assetId, client }).send();
-};
-
-const listPropertiesSiteWiseAssetQuery = async (
-  client: IoTSiteWiseClient,
-  siteWiseQuery: Partial<SiteWiseAssetQuery & SiteWisePropertyAliasQuery>
-) =>
-  Promise.all(
-    siteWiseQuery.assets?.map(({ assetId }: { assetId: string }) =>
-      listAssetProperties(client, assetId)
-    ) ?? []
-  );
-
-const newMapPropertySummary = ({
-  id,
-  unit,
-  alias,
-  path,
-  dataType,
-}: AssetModelPropertySummary & AssetPropertySummary): PropertySummary => ({
-  propertyId: id,
-  name: path?.[path.length - 1].name,
-  unit,
-  dataType,
-  alias,
-});
-
-const mapListAssetPropertiesToAssetSummary = (
-  n: (AssetModelPropertySummary & AssetPropertySummary)[],
-  assetId: string,
-  assetName: string | undefined
-): AssetSummary => {
-  return {
-    assetId: assetId,
-    assetName: assetName,
-    properties: n.map(newMapPropertySummary) ?? [],
-    alarms: [],
-  };
-};
-
 export const useListAssetPropertiesMapQuery = (
   siteWiseQuery:
     | Partial<SiteWiseAssetQuery & SiteWisePropertyAliasQuery>
     | undefined
 ) => {
   const { iotSiteWiseClient } = useClients();
-  const fetchSiteWiseAssetQueryDescription = async () => {
-    if (!iotSiteWiseClient || !siteWiseQuery) return [];
-    return listPropertiesSiteWiseAssetQuery(iotSiteWiseClient, siteWiseQuery);
-  };
+  const queryKey = createListAssetPropertiesMapCacheKey(siteWiseQuery);
+  const queryFn = createFetchSiteWiseAssetQueryDescription(
+    iotSiteWiseClient,
+    siteWiseQuery
+  );
 
   return useQuery({
-    queryKey: [
-      ...ASSET_DESCRIPTION_QUERY_KEY,
-      'assetDescriptionsMap',
-      ...(siteWiseQuery?.assets?.map((a) => a.assetId) ?? []),
-    ],
-    queryFn: () => fetchSiteWiseAssetQueryDescription(),
-    select: (data) =>
-      data?.reduce<Record<string, AssetSummary>>((acc, n) => {
-        const assetId = n.at(0)?.path?.at(0)?.id;
-        if (assetId) {
-          const assetName = n.at(0)?.path?.at(0)?.name;
-          acc[assetId] = mapListAssetPropertiesToAssetSummary(
-            n,
-            assetId,
-            assetName
-          );
-        }
-        return acc;
-      }, {}) ?? {},
-  });
-};
-
-//dead code?? Please let me know if we should remove in PR review
-
-export const useAssetDescriptionQuery = (assetId: string | undefined) => {
-  const { iotSiteWiseClient } = useClients();
-
-  const fetchAssetDescription = async () => {
-    if (!iotSiteWiseClient || !assetId) return;
-
-    return describeAsset(iotSiteWiseClient, assetId);
-  };
-
-  return useQuery({
-    queryKey: [...ASSET_DESCRIPTION_QUERY_KEY, assetId],
-    queryFn: () => fetchAssetDescription(),
-    select: (data) => (data ? mapAssetDescriptionToAssetSummary(data) : data),
-    enabled: !!assetId,
+    queryKey,
+    queryFn,
+    select: selectListAssetPropertiesMap,
   });
 };
